+++ examples/BikeNYC/L4-C3-P1-T1/config.ini
@@ -1,0 +1,24 @@
+[dataset]
+data_files = ["./datasets/BikeNYC/NYC14_M16x8_T60_NewEnd.h5"]
+holiday_file = "./datasets/BikeNYC/NY_Holiday.txt"
+meteorol_file = ""
+T = 24
+len_closeness = 3
+len_period = 1
+len_trend = 1
+period_interval = 1
+trend_interval = 7
+; 24 * 28
+len_test = 672
+use_meta = true
+
+[model]
+nb_flow = 2
+map_height = 8
+map_width = 16
+nb_residual_unit = 4
+
+[learning]
+epochs = 100
+batch_size = 32
+learning_rate = 0.0002
