--- examples/L12-C3-P1-T1/log.txt
@@ -1,100 +1,0 @@
-[2022-08-21 10:44:33] :stresnet.utils.logger: [INFO] loss: 37.11479949951172
-[2022-08-21 10:45:55] :stresnet.utils.logger: [INFO] loss: 27.488037109375
-[2022-08-21 10:47:18] :stresnet.utils.logger: [INFO] loss: 23.65913963317871
-[2022-08-21 10:48:40] :stresnet.utils.logger: [INFO] loss: 22.130855560302734
-[2022-08-21 10:50:03] :stresnet.utils.logger: [INFO] loss: 19.937320709228516
-[2022-08-21 10:51:26] :stresnet.utils.logger: [INFO] loss: 19.964628219604492
-[2022-08-21 10:52:48] :stresnet.utils.logger: [INFO] loss: 19.59438705444336
-[2022-08-21 10:54:11] :stresnet.utils.logger: [INFO] loss: 19.860095977783203
-[2022-08-21 10:55:34] :stresnet.utils.logger: [INFO] loss: 19.129680633544922
-[2022-08-21 10:56:56] :stresnet.utils.logger: [INFO] loss: 19.791980743408203
-[2022-08-21 10:58:19] :stresnet.utils.logger: [INFO] loss: 19.75201416015625
-[2022-08-21 10:59:41] :stresnet.utils.logger: [INFO] loss: 21.019779205322266
-[2022-08-21 11:01:04] :stresnet.utils.logger: [INFO] loss: 20.41977882385254
-[2022-08-21 11:02:26] :stresnet.utils.logger: [INFO] loss: 20.417787551879883
-[2022-08-21 11:03:49] :stresnet.utils.logger: [INFO] loss: 20.519983291625977
-[2022-08-21 11:05:11] :stresnet.utils.logger: [INFO] loss: 20.497066497802734
-[2022-08-21 11:06:33] :stresnet.utils.logger: [INFO] loss: 19.148469924926758
-[2022-08-21 11:07:56] :stresnet.utils.logger: [INFO] loss: 19.602432250976562
-[2022-08-21 11:09:18] :stresnet.utils.logger: [INFO] loss: 18.830509185791016
-[2022-08-21 11:10:41] :stresnet.utils.logger: [INFO] loss: 18.37575912475586
-[2022-08-21 11:12:03] :stresnet.utils.logger: [INFO] loss: 18.56191062927246
-[2022-08-21 11:13:26] :stresnet.utils.logger: [INFO] loss: 17.913179397583008
-[2022-08-21 11:14:48] :stresnet.utils.logger: [INFO] loss: 17.623626708984375
-[2022-08-21 11:16:10] :stresnet.utils.logger: [INFO] loss: 17.020660400390625
-[2022-08-21 11:17:33] :stresnet.utils.logger: [INFO] loss: 16.037141799926758
-[2022-08-21 11:18:55] :stresnet.utils.logger: [INFO] loss: 17.186735153198242
-[2022-08-21 11:20:18] :stresnet.utils.logger: [INFO] loss: 17.627967834472656
-[2022-08-21 11:21:40] :stresnet.utils.logger: [INFO] loss: 17.726806640625
-[2022-08-21 11:23:02] :stresnet.utils.logger: [INFO] loss: 17.930496215820312
-[2022-08-21 11:24:25] :stresnet.utils.logger: [INFO] loss: 17.79239845275879
-[2022-08-21 11:25:47] :stresnet.utils.logger: [INFO] loss: 17.98787498474121
-[2022-08-21 11:27:10] :stresnet.utils.logger: [INFO] loss: 18.53011131286621
-[2022-08-21 11:28:32] :stresnet.utils.logger: [INFO] loss: 17.458629608154297
-[2022-08-21 11:29:54] :stresnet.utils.logger: [INFO] loss: 16.7348575592041
-[2022-08-21 11:31:17] :stresnet.utils.logger: [INFO] loss: 16.28895378112793
-[2022-08-21 11:32:39] :stresnet.utils.logger: [INFO] loss: 16.238468170166016
-[2022-08-21 11:34:01] :stresnet.utils.logger: [INFO] loss: 16.52086639404297
-[2022-08-21 11:35:24] :stresnet.utils.logger: [INFO] loss: 16.766075134277344
-[2022-08-21 11:36:46] :stresnet.utils.logger: [INFO] loss: 17.43662452697754
-[2022-08-21 11:38:09] :stresnet.utils.logger: [INFO] loss: 17.226585388183594
-[2022-08-21 11:39:31] :stresnet.utils.logger: [INFO] loss: 17.060680389404297
-[2022-08-21 11:40:53] :stresnet.utils.logger: [INFO] loss: 17.152318954467773
-[2022-08-21 11:42:16] :stresnet.utils.logger: [INFO] loss: 16.413427352905273
-[2022-08-21 11:43:38] :stresnet.utils.logger: [INFO] loss: 16.192806243896484
-[2022-08-21 11:45:01] :stresnet.utils.logger: [INFO] loss: 16.713159561157227
-[2022-08-21 11:46:23] :stresnet.utils.logger: [INFO] loss: 17.23483657836914
-[2022-08-21 11:47:45] :stresnet.utils.logger: [INFO] loss: 18.347843170166016
-[2022-08-21 11:49:08] :stresnet.utils.logger: [INFO] loss: 18.00539207458496
-[2022-08-21 11:50:30] :stresnet.utils.logger: [INFO] loss: 18.37088966369629
-[2022-08-21 11:51:53] :stresnet.utils.logger: [INFO] loss: 18.77469253540039
-[2022-08-21 11:53:15] :stresnet.utils.logger: [INFO] loss: 18.431705474853516
-[2022-08-21 11:54:37] :stresnet.utils.logger: [INFO] loss: 20.23248863220215
-[2022-08-21 11:56:00] :stresnet.utils.logger: [INFO] loss: 17.593259811401367
-[2022-08-21 11:57:22] :stresnet.utils.logger: [INFO] loss: 16.719106674194336
-[2022-08-21 11:58:45] :stresnet.utils.logger: [INFO] loss: 19.417146682739258
-[2022-08-21 12:00:07] :stresnet.utils.logger: [INFO] loss: 18.334077835083008
-[2022-08-21 12:01:29] :stresnet.utils.logger: [INFO] loss: 18.24327278137207
-[2022-08-21 12:02:52] :stresnet.utils.logger: [INFO] loss: 17.86640739440918
-[2022-08-21 12:04:14] :stresnet.utils.logger: [INFO] loss: 19.08664321899414
-[2022-08-21 12:05:36] :stresnet.utils.logger: [INFO] loss: 17.05435562133789
-[2022-08-21 12:06:59] :stresnet.utils.logger: [INFO] loss: 13.649744987487793
-[2022-08-21 12:08:21] :stresnet.utils.logger: [INFO] loss: 12.564355850219727
-[2022-08-21 12:09:44] :stresnet.utils.logger: [INFO] loss: 12.679534912109375
-[2022-08-21 12:11:06] :stresnet.utils.logger: [INFO] loss: 13.30785083770752
-[2022-08-21 12:12:28] :stresnet.utils.logger: [INFO] loss: 14.999470710754395
-[2022-08-21 12:13:51] :stresnet.utils.logger: [INFO] loss: 15.487525939941406
-[2022-08-21 12:15:13] :stresnet.utils.logger: [INFO] loss: 15.737791061401367
-[2022-08-21 12:16:36] :stresnet.utils.logger: [INFO] loss: 15.633357048034668
-[2022-08-21 12:17:58] :stresnet.utils.logger: [INFO] loss: 14.443073272705078
-[2022-08-21 12:19:21] :stresnet.utils.logger: [INFO] loss: 14.160954475402832
-[2022-08-21 12:20:43] :stresnet.utils.logger: [INFO] loss: 13.783675193786621
-[2022-08-21 12:22:05] :stresnet.utils.logger: [INFO] loss: 13.986112594604492
-[2022-08-21 12:23:28] :stresnet.utils.logger: [INFO] loss: 13.336237907409668
-[2022-08-21 12:24:50] :stresnet.utils.logger: [INFO] loss: 13.103713035583496
-[2022-08-21 12:26:13] :stresnet.utils.logger: [INFO] loss: 12.978813171386719
-[2022-08-21 12:27:35] :stresnet.utils.logger: [INFO] loss: 12.796905517578125
-[2022-08-21 12:28:57] :stresnet.utils.logger: [INFO] loss: 13.411384582519531
-[2022-08-21 12:30:20] :stresnet.utils.logger: [INFO] loss: 14.054084777832031
-[2022-08-21 12:31:42] :stresnet.utils.logger: [INFO] loss: 13.302135467529297
-[2022-08-21 12:33:05] :stresnet.utils.logger: [INFO] loss: 13.498299598693848
-[2022-08-21 12:34:27] :stresnet.utils.logger: [INFO] loss: 13.204378128051758
-[2022-08-21 12:35:49] :stresnet.utils.logger: [INFO] loss: 13.364341735839844
-[2022-08-21 12:37:12] :stresnet.utils.logger: [INFO] loss: 12.199922561645508
-[2022-08-21 12:38:34] :stresnet.utils.logger: [INFO] loss: 13.384063720703125
-[2022-08-21 12:39:57] :stresnet.utils.logger: [INFO] loss: 12.946840286254883
-[2022-08-21 12:41:19] :stresnet.utils.logger: [INFO] loss: 13.454145431518555
-[2022-08-21 12:42:42] :stresnet.utils.logger: [INFO] loss: 12.326292037963867
-[2022-08-21 12:44:04] :stresnet.utils.logger: [INFO] loss: 12.603873252868652
-[2022-08-21 12:45:27] :stresnet.utils.logger: [INFO] loss: 12.429892539978027
-[2022-08-21 12:46:49] :stresnet.utils.logger: [INFO] loss: 13.120804786682129
-[2022-08-21 12:48:12] :stresnet.utils.logger: [INFO] loss: 12.039324760437012
-[2022-08-21 12:49:34] :stresnet.utils.logger: [INFO] loss: 12.463050842285156
-[2022-08-21 12:50:56] :stresnet.utils.logger: [INFO] loss: 11.990405082702637
-[2022-08-21 12:52:19] :stresnet.utils.logger: [INFO] loss: 12.428117752075195
-[2022-08-21 12:53:41] :stresnet.utils.logger: [INFO] loss: 15.554362297058105
-[2022-08-21 12:55:04] :stresnet.utils.logger: [INFO] loss: 12.467930793762207
-[2022-08-21 12:56:26] :stresnet.utils.logger: [INFO] loss: 11.717819213867188
-[2022-08-21 12:57:49] :stresnet.utils.logger: [INFO] loss: 12.74034595489502
-[2022-08-21 12:59:11] :stresnet.utils.logger: [INFO] loss: 11.979413032531738
-[2022-08-21 13:00:33] :stresnet.utils.logger: [INFO] loss: 13.533787727355957
