--- examples/L4-C3-P1-T1/log.txt
@@ -1,100 +1,0 @@
-[2022-08-21 09:54:54] :stresnet.utils.logger: [INFO] loss: 35.77045440673828
-[2022-08-21 09:55:23] :stresnet.utils.logger: [INFO] loss: 28.707412719726562
-[2022-08-21 09:55:52] :stresnet.utils.logger: [INFO] loss: 27.01711082458496
-[2022-08-21 09:56:20] :stresnet.utils.logger: [INFO] loss: 22.60215950012207
-[2022-08-21 09:56:49] :stresnet.utils.logger: [INFO] loss: 20.654861450195312
-[2022-08-21 09:57:18] :stresnet.utils.logger: [INFO] loss: 20.087467193603516
-[2022-08-21 09:57:46] :stresnet.utils.logger: [INFO] loss: 19.14936637878418
-[2022-08-21 09:58:15] :stresnet.utils.logger: [INFO] loss: 18.702852249145508
-[2022-08-21 09:58:43] :stresnet.utils.logger: [INFO] loss: 18.64673614501953
-[2022-08-21 09:59:12] :stresnet.utils.logger: [INFO] loss: 17.72690200805664
-[2022-08-21 09:59:41] :stresnet.utils.logger: [INFO] loss: 17.07025146484375
-[2022-08-21 10:00:09] :stresnet.utils.logger: [INFO] loss: 17.372840881347656
-[2022-08-21 10:00:38] :stresnet.utils.logger: [INFO] loss: 17.65348243713379
-[2022-08-21 10:01:07] :stresnet.utils.logger: [INFO] loss: 17.702890396118164
-[2022-08-21 10:01:35] :stresnet.utils.logger: [INFO] loss: 18.000728607177734
-[2022-08-21 10:02:04] :stresnet.utils.logger: [INFO] loss: 18.434804916381836
-[2022-08-21 10:02:33] :stresnet.utils.logger: [INFO] loss: 16.944005966186523
-[2022-08-21 10:03:01] :stresnet.utils.logger: [INFO] loss: 17.199447631835938
-[2022-08-21 10:03:30] :stresnet.utils.logger: [INFO] loss: 17.216201782226562
-[2022-08-21 10:03:59] :stresnet.utils.logger: [INFO] loss: 16.126012802124023
-[2022-08-21 10:04:27] :stresnet.utils.logger: [INFO] loss: 15.378118515014648
-[2022-08-21 10:04:56] :stresnet.utils.logger: [INFO] loss: 15.126961708068848
-[2022-08-21 10:05:25] :stresnet.utils.logger: [INFO] loss: 14.814445495605469
-[2022-08-21 10:05:53] :stresnet.utils.logger: [INFO] loss: 14.59094524383545
-[2022-08-21 10:06:22] :stresnet.utils.logger: [INFO] loss: 14.48259162902832
-[2022-08-21 10:06:51] :stresnet.utils.logger: [INFO] loss: 14.434913635253906
-[2022-08-21 10:07:20] :stresnet.utils.logger: [INFO] loss: 14.404976844787598
-[2022-08-21 10:07:48] :stresnet.utils.logger: [INFO] loss: 14.2808198928833
-[2022-08-21 10:08:17] :stresnet.utils.logger: [INFO] loss: 13.848167419433594
-[2022-08-21 10:08:46] :stresnet.utils.logger: [INFO] loss: 13.379862785339355
-[2022-08-21 10:09:14] :stresnet.utils.logger: [INFO] loss: 13.04993724822998
-[2022-08-21 10:09:43] :stresnet.utils.logger: [INFO] loss: 12.78166675567627
-[2022-08-21 10:10:12] :stresnet.utils.logger: [INFO] loss: 12.520133018493652
-[2022-08-21 10:10:41] :stresnet.utils.logger: [INFO] loss: 12.392400741577148
-[2022-08-21 10:11:09] :stresnet.utils.logger: [INFO] loss: 12.44774055480957
-[2022-08-21 10:11:38] :stresnet.utils.logger: [INFO] loss: 12.383842468261719
-[2022-08-21 10:12:07] :stresnet.utils.logger: [INFO] loss: 12.29627799987793
-[2022-08-21 10:12:36] :stresnet.utils.logger: [INFO] loss: 12.208415985107422
-[2022-08-21 10:13:04] :stresnet.utils.logger: [INFO] loss: 12.15043830871582
-[2022-08-21 10:13:33] :stresnet.utils.logger: [INFO] loss: 12.082810401916504
-[2022-08-21 10:14:02] :stresnet.utils.logger: [INFO] loss: 12.037074089050293
-[2022-08-21 10:14:31] :stresnet.utils.logger: [INFO] loss: 12.09093189239502
-[2022-08-21 10:14:59] :stresnet.utils.logger: [INFO] loss: 12.114747047424316
-[2022-08-21 10:15:28] :stresnet.utils.logger: [INFO] loss: 12.239313125610352
-[2022-08-21 10:15:57] :stresnet.utils.logger: [INFO] loss: 12.273640632629395
-[2022-08-21 10:16:26] :stresnet.utils.logger: [INFO] loss: 12.330060958862305
-[2022-08-21 10:16:54] :stresnet.utils.logger: [INFO] loss: 12.26408863067627
-[2022-08-21 10:17:23] :stresnet.utils.logger: [INFO] loss: 12.320551872253418
-[2022-08-21 10:17:52] :stresnet.utils.logger: [INFO] loss: 12.472509384155273
-[2022-08-21 10:18:20] :stresnet.utils.logger: [INFO] loss: 12.556892395019531
-[2022-08-21 10:18:49] :stresnet.utils.logger: [INFO] loss: 12.786518096923828
-[2022-08-21 10:19:18] :stresnet.utils.logger: [INFO] loss: 12.885076522827148
-[2022-08-21 10:19:47] :stresnet.utils.logger: [INFO] loss: 12.995993614196777
-[2022-08-21 10:20:15] :stresnet.utils.logger: [INFO] loss: 13.123455047607422
-[2022-08-21 10:20:44] :stresnet.utils.logger: [INFO] loss: 13.290810585021973
-[2022-08-21 10:21:13] :stresnet.utils.logger: [INFO] loss: 13.289588928222656
-[2022-08-21 10:21:42] :stresnet.utils.logger: [INFO] loss: 13.405769348144531
-[2022-08-21 10:22:10] :stresnet.utils.logger: [INFO] loss: 13.644505500793457
-[2022-08-21 10:22:39] :stresnet.utils.logger: [INFO] loss: 13.865800857543945
-[2022-08-21 10:23:08] :stresnet.utils.logger: [INFO] loss: 13.958684921264648
-[2022-08-21 10:23:37] :stresnet.utils.logger: [INFO] loss: 14.377391815185547
-[2022-08-21 10:24:05] :stresnet.utils.logger: [INFO] loss: 14.44996166229248
-[2022-08-21 10:24:34] :stresnet.utils.logger: [INFO] loss: 14.41618824005127
-[2022-08-21 10:25:03] :stresnet.utils.logger: [INFO] loss: 14.474151611328125
-[2022-08-21 10:25:32] :stresnet.utils.logger: [INFO] loss: 14.759068489074707
-[2022-08-21 10:26:00] :stresnet.utils.logger: [INFO] loss: 14.703102111816406
-[2022-08-21 10:26:29] :stresnet.utils.logger: [INFO] loss: 14.722557067871094
-[2022-08-21 10:26:58] :stresnet.utils.logger: [INFO] loss: 14.49024486541748
-[2022-08-21 10:27:27] :stresnet.utils.logger: [INFO] loss: 14.344331741333008
-[2022-08-21 10:27:55] :stresnet.utils.logger: [INFO] loss: 14.144514083862305
-[2022-08-21 10:28:24] :stresnet.utils.logger: [INFO] loss: 13.936334609985352
-[2022-08-21 10:28:53] :stresnet.utils.logger: [INFO] loss: 13.84277629852295
-[2022-08-21 10:29:21] :stresnet.utils.logger: [INFO] loss: 13.769254684448242
-[2022-08-21 10:29:50] :stresnet.utils.logger: [INFO] loss: 13.501659393310547
-[2022-08-21 10:30:19] :stresnet.utils.logger: [INFO] loss: 13.396739959716797
-[2022-08-21 10:30:48] :stresnet.utils.logger: [INFO] loss: 13.19813346862793
-[2022-08-21 10:31:16] :stresnet.utils.logger: [INFO] loss: 13.051605224609375
-[2022-08-21 10:31:45] :stresnet.utils.logger: [INFO] loss: 12.858609199523926
-[2022-08-21 10:32:14] :stresnet.utils.logger: [INFO] loss: 12.841394424438477
-[2022-08-21 10:32:43] :stresnet.utils.logger: [INFO] loss: 12.799070358276367
-[2022-08-21 10:33:11] :stresnet.utils.logger: [INFO] loss: 12.766130447387695
-[2022-08-21 10:33:40] :stresnet.utils.logger: [INFO] loss: 12.537237167358398
-[2022-08-21 10:34:09] :stresnet.utils.logger: [INFO] loss: 12.533125877380371
-[2022-08-21 10:34:38] :stresnet.utils.logger: [INFO] loss: 12.524015426635742
-[2022-08-21 10:35:06] :stresnet.utils.logger: [INFO] loss: 12.5184965133667
-[2022-08-21 10:35:35] :stresnet.utils.logger: [INFO] loss: 12.548380851745605
-[2022-08-21 10:36:04] :stresnet.utils.logger: [INFO] loss: 12.615859031677246
-[2022-08-21 10:36:33] :stresnet.utils.logger: [INFO] loss: 12.741860389709473
-[2022-08-21 10:37:01] :stresnet.utils.logger: [INFO] loss: 12.766168594360352
-[2022-08-21 10:37:30] :stresnet.utils.logger: [INFO] loss: 12.790071487426758
-[2022-08-21 10:37:59] :stresnet.utils.logger: [INFO] loss: 12.814591407775879
-[2022-08-21 10:38:28] :stresnet.utils.logger: [INFO] loss: 12.901381492614746
-[2022-08-21 10:38:56] :stresnet.utils.logger: [INFO] loss: 12.896326065063477
-[2022-08-21 10:39:25] :stresnet.utils.logger: [INFO] loss: 12.894773483276367
-[2022-08-21 10:39:54] :stresnet.utils.logger: [INFO] loss: 12.877893447875977
-[2022-08-21 10:40:23] :stresnet.utils.logger: [INFO] loss: 12.91377067565918
-[2022-08-21 10:40:51] :stresnet.utils.logger: [INFO] loss: 12.807878494262695
-[2022-08-21 10:41:20] :stresnet.utils.logger: [INFO] loss: 12.6741943359375
-[2022-08-21 10:41:49] :stresnet.utils.logger: [INFO] loss: 12.649510383605957
-[2022-08-21 10:42:18] :stresnet.utils.logger: [INFO] loss: 12.726140975952148
