--- examples/L4-C3-P3-T3/log.txt
@@ -1,100 +1,0 @@
-[2022-08-25 08:58:02] :stresnet.utils.logger: [INFO] loss: 57.52571105957031
-[2022-08-25 08:58:16] :stresnet.utils.logger: [INFO] loss: 42.388919830322266
-[2022-08-25 08:58:30] :stresnet.utils.logger: [INFO] loss: 40.639156341552734
-[2022-08-25 08:58:44] :stresnet.utils.logger: [INFO] loss: 35.937137603759766
-[2022-08-25 08:58:58] :stresnet.utils.logger: [INFO] loss: 33.47725296020508
-[2022-08-25 08:59:12] :stresnet.utils.logger: [INFO] loss: 32.192020416259766
-[2022-08-25 08:59:27] :stresnet.utils.logger: [INFO] loss: 30.305503845214844
-[2022-08-25 08:59:41] :stresnet.utils.logger: [INFO] loss: 29.122037887573242
-[2022-08-25 08:59:55] :stresnet.utils.logger: [INFO] loss: 27.910484313964844
-[2022-08-25 09:00:09] :stresnet.utils.logger: [INFO] loss: 26.983657836914062
-[2022-08-25 09:00:23] :stresnet.utils.logger: [INFO] loss: 25.906145095825195
-[2022-08-25 09:00:37] :stresnet.utils.logger: [INFO] loss: 24.176191329956055
-[2022-08-25 09:00:51] :stresnet.utils.logger: [INFO] loss: 22.627277374267578
-[2022-08-25 09:01:05] :stresnet.utils.logger: [INFO] loss: 21.576658248901367
-[2022-08-25 09:01:20] :stresnet.utils.logger: [INFO] loss: 20.9113712310791
-[2022-08-25 09:01:34] :stresnet.utils.logger: [INFO] loss: 20.050228118896484
-[2022-08-25 09:01:48] :stresnet.utils.logger: [INFO] loss: 19.482864379882812
-[2022-08-25 09:02:02] :stresnet.utils.logger: [INFO] loss: 18.958251953125
-[2022-08-25 09:02:16] :stresnet.utils.logger: [INFO] loss: 18.459280014038086
-[2022-08-25 09:02:30] :stresnet.utils.logger: [INFO] loss: 18.111154556274414
-[2022-08-25 09:02:45] :stresnet.utils.logger: [INFO] loss: 17.6739444732666
-[2022-08-25 09:02:59] :stresnet.utils.logger: [INFO] loss: 17.285303115844727
-[2022-08-25 09:03:13] :stresnet.utils.logger: [INFO] loss: 17.05959701538086
-[2022-08-25 09:03:27] :stresnet.utils.logger: [INFO] loss: 16.886457443237305
-[2022-08-25 09:03:41] :stresnet.utils.logger: [INFO] loss: 16.813983917236328
-[2022-08-25 09:03:55] :stresnet.utils.logger: [INFO] loss: 16.555604934692383
-[2022-08-25 09:04:10] :stresnet.utils.logger: [INFO] loss: 16.501924514770508
-[2022-08-25 09:04:24] :stresnet.utils.logger: [INFO] loss: 16.1597957611084
-[2022-08-25 09:04:38] :stresnet.utils.logger: [INFO] loss: 16.00833511352539
-[2022-08-25 09:04:52] :stresnet.utils.logger: [INFO] loss: 15.9572172164917
-[2022-08-25 09:05:06] :stresnet.utils.logger: [INFO] loss: 16.153865814208984
-[2022-08-25 09:05:20] :stresnet.utils.logger: [INFO] loss: 16.358089447021484
-[2022-08-25 09:05:34] :stresnet.utils.logger: [INFO] loss: 16.549684524536133
-[2022-08-25 09:05:49] :stresnet.utils.logger: [INFO] loss: 16.8038330078125
-[2022-08-25 09:06:03] :stresnet.utils.logger: [INFO] loss: 16.568696975708008
-[2022-08-25 09:06:17] :stresnet.utils.logger: [INFO] loss: 16.246009826660156
-[2022-08-25 09:06:31] :stresnet.utils.logger: [INFO] loss: 15.939361572265625
-[2022-08-25 09:06:45] :stresnet.utils.logger: [INFO] loss: 15.723920822143555
-[2022-08-25 09:06:59] :stresnet.utils.logger: [INFO] loss: 15.636558532714844
-[2022-08-25 09:07:14] :stresnet.utils.logger: [INFO] loss: 15.379005432128906
-[2022-08-25 09:07:28] :stresnet.utils.logger: [INFO] loss: 15.221385955810547
-[2022-08-25 09:07:42] :stresnet.utils.logger: [INFO] loss: 15.127156257629395
-[2022-08-25 09:07:56] :stresnet.utils.logger: [INFO] loss: 15.0296630859375
-[2022-08-25 09:08:10] :stresnet.utils.logger: [INFO] loss: 14.970802307128906
-[2022-08-25 09:08:24] :stresnet.utils.logger: [INFO] loss: 14.849346160888672
-[2022-08-25 09:08:39] :stresnet.utils.logger: [INFO] loss: 14.802229881286621
-[2022-08-25 09:08:53] :stresnet.utils.logger: [INFO] loss: 14.740875244140625
-[2022-08-25 09:09:07] :stresnet.utils.logger: [INFO] loss: 14.754594802856445
-[2022-08-25 09:09:21] :stresnet.utils.logger: [INFO] loss: 14.687627792358398
-[2022-08-25 09:09:35] :stresnet.utils.logger: [INFO] loss: 14.707083702087402
-[2022-08-25 09:09:49] :stresnet.utils.logger: [INFO] loss: 14.840249061584473
-[2022-08-25 09:10:03] :stresnet.utils.logger: [INFO] loss: 15.116206169128418
-[2022-08-25 09:10:18] :stresnet.utils.logger: [INFO] loss: 15.245081901550293
-[2022-08-25 09:10:32] :stresnet.utils.logger: [INFO] loss: 14.893082618713379
-[2022-08-25 09:10:46] :stresnet.utils.logger: [INFO] loss: 14.470191955566406
-[2022-08-25 09:11:00] :stresnet.utils.logger: [INFO] loss: 14.299871444702148
-[2022-08-25 09:11:14] :stresnet.utils.logger: [INFO] loss: 14.17506217956543
-[2022-08-25 09:11:28] :stresnet.utils.logger: [INFO] loss: 14.162458419799805
-[2022-08-25 09:11:43] :stresnet.utils.logger: [INFO] loss: 14.214510917663574
-[2022-08-25 09:11:57] :stresnet.utils.logger: [INFO] loss: 14.192131042480469
-[2022-08-25 09:12:11] :stresnet.utils.logger: [INFO] loss: 14.512112617492676
-[2022-08-25 09:12:25] :stresnet.utils.logger: [INFO] loss: 15.176301956176758
-[2022-08-25 09:12:39] :stresnet.utils.logger: [INFO] loss: 19.302093505859375
-[2022-08-25 09:12:53] :stresnet.utils.logger: [INFO] loss: 22.94656753540039
-[2022-08-25 09:13:07] :stresnet.utils.logger: [INFO] loss: 17.175704956054688
-[2022-08-25 09:13:22] :stresnet.utils.logger: [INFO] loss: 16.872793197631836
-[2022-08-25 09:13:36] :stresnet.utils.logger: [INFO] loss: 17.778423309326172
-[2022-08-25 09:13:50] :stresnet.utils.logger: [INFO] loss: 18.330223083496094
-[2022-08-25 09:14:04] :stresnet.utils.logger: [INFO] loss: 18.49073028564453
-[2022-08-25 09:14:18] :stresnet.utils.logger: [INFO] loss: 18.18414878845215
-[2022-08-25 09:14:32] :stresnet.utils.logger: [INFO] loss: 17.860336303710938
-[2022-08-25 09:14:47] :stresnet.utils.logger: [INFO] loss: 17.50272560119629
-[2022-08-25 09:15:01] :stresnet.utils.logger: [INFO] loss: 17.231027603149414
-[2022-08-25 09:15:15] :stresnet.utils.logger: [INFO] loss: 17.009231567382812
-[2022-08-25 09:15:29] :stresnet.utils.logger: [INFO] loss: 16.705976486206055
-[2022-08-25 09:15:43] :stresnet.utils.logger: [INFO] loss: 16.425277709960938
-[2022-08-25 09:15:57] :stresnet.utils.logger: [INFO] loss: 16.089366912841797
-[2022-08-25 09:16:11] :stresnet.utils.logger: [INFO] loss: 15.822751998901367
-[2022-08-25 09:16:26] :stresnet.utils.logger: [INFO] loss: 15.53044319152832
-[2022-08-25 09:16:40] :stresnet.utils.logger: [INFO] loss: 15.324498176574707
-[2022-08-25 09:16:54] :stresnet.utils.logger: [INFO] loss: 15.208382606506348
-[2022-08-25 09:17:08] :stresnet.utils.logger: [INFO] loss: 15.097630500793457
-[2022-08-25 09:17:22] :stresnet.utils.logger: [INFO] loss: 15.081092834472656
-[2022-08-25 09:17:36] :stresnet.utils.logger: [INFO] loss: 15.089741706848145
-[2022-08-25 09:17:50] :stresnet.utils.logger: [INFO] loss: 15.14891242980957
-[2022-08-25 09:18:05] :stresnet.utils.logger: [INFO] loss: 15.204710006713867
-[2022-08-25 09:18:19] :stresnet.utils.logger: [INFO] loss: 15.185945510864258
-[2022-08-25 09:18:33] :stresnet.utils.logger: [INFO] loss: 15.232102394104004
-[2022-08-25 09:18:47] :stresnet.utils.logger: [INFO] loss: 15.181532859802246
-[2022-08-25 09:19:01] :stresnet.utils.logger: [INFO] loss: 15.124381065368652
-[2022-08-25 09:19:15] :stresnet.utils.logger: [INFO] loss: 15.151412010192871
-[2022-08-25 09:19:30] :stresnet.utils.logger: [INFO] loss: 15.260733604431152
-[2022-08-25 09:19:44] :stresnet.utils.logger: [INFO] loss: 15.439797401428223
-[2022-08-25 09:19:58] :stresnet.utils.logger: [INFO] loss: 15.472892761230469
-[2022-08-25 09:20:12] :stresnet.utils.logger: [INFO] loss: 15.620539665222168
-[2022-08-25 09:20:26] :stresnet.utils.logger: [INFO] loss: 15.76710319519043
-[2022-08-25 09:20:40] :stresnet.utils.logger: [INFO] loss: 16.09638214111328
-[2022-08-25 09:20:54] :stresnet.utils.logger: [INFO] loss: 16.16303062438965
-[2022-08-25 09:21:09] :stresnet.utils.logger: [INFO] loss: 16.33378028869629
-[2022-08-25 09:21:23] :stresnet.utils.logger: [INFO] loss: 16.368486404418945
