--- examples/L6-C3-P1-T1/log.txt
@@ -1,100 +1,0 @@
-[2022-08-21 13:12:23] :stresnet.utils.logger: [INFO] loss: 33.702030181884766
-[2022-08-21 13:13:05] :stresnet.utils.logger: [INFO] loss: 26.216197967529297
-[2022-08-21 13:13:47] :stresnet.utils.logger: [INFO] loss: 23.615345001220703
-[2022-08-21 13:14:28] :stresnet.utils.logger: [INFO] loss: 21.281902313232422
-[2022-08-21 13:15:10] :stresnet.utils.logger: [INFO] loss: 19.92637825012207
-[2022-08-21 13:15:52] :stresnet.utils.logger: [INFO] loss: 19.405866622924805
-[2022-08-21 13:16:34] :stresnet.utils.logger: [INFO] loss: 19.204513549804688
-[2022-08-21 13:17:15] :stresnet.utils.logger: [INFO] loss: 18.871135711669922
-[2022-08-21 13:17:57] :stresnet.utils.logger: [INFO] loss: 18.391048431396484
-[2022-08-21 13:18:39] :stresnet.utils.logger: [INFO] loss: 18.970800399780273
-[2022-08-21 13:19:21] :stresnet.utils.logger: [INFO] loss: 18.255605697631836
-[2022-08-21 13:20:02] :stresnet.utils.logger: [INFO] loss: 17.68404769897461
-[2022-08-21 13:20:44] :stresnet.utils.logger: [INFO] loss: 17.419206619262695
-[2022-08-21 13:21:26] :stresnet.utils.logger: [INFO] loss: 17.952938079833984
-[2022-08-21 13:22:08] :stresnet.utils.logger: [INFO] loss: 18.411869049072266
-[2022-08-21 13:22:49] :stresnet.utils.logger: [INFO] loss: 18.44534683227539
-[2022-08-21 13:23:31] :stresnet.utils.logger: [INFO] loss: 18.788816452026367
-[2022-08-21 13:24:13] :stresnet.utils.logger: [INFO] loss: 20.133852005004883
-[2022-08-21 13:24:54] :stresnet.utils.logger: [INFO] loss: 20.804710388183594
-[2022-08-21 13:25:36] :stresnet.utils.logger: [INFO] loss: 21.062210083007812
-[2022-08-21 13:26:18] :stresnet.utils.logger: [INFO] loss: 19.698461532592773
-[2022-08-21 13:27:00] :stresnet.utils.logger: [INFO] loss: 18.98756980895996
-[2022-08-21 13:27:41] :stresnet.utils.logger: [INFO] loss: 18.075138092041016
-[2022-08-21 13:28:23] :stresnet.utils.logger: [INFO] loss: 17.957355499267578
-[2022-08-21 13:29:05] :stresnet.utils.logger: [INFO] loss: 18.13149642944336
-[2022-08-21 13:29:47] :stresnet.utils.logger: [INFO] loss: 18.385761260986328
-[2022-08-21 13:30:28] :stresnet.utils.logger: [INFO] loss: 17.954452514648438
-[2022-08-21 13:31:10] :stresnet.utils.logger: [INFO] loss: 17.255504608154297
-[2022-08-21 13:31:52] :stresnet.utils.logger: [INFO] loss: 17.451358795166016
-[2022-08-21 13:32:33] :stresnet.utils.logger: [INFO] loss: 17.81602668762207
-[2022-08-21 13:33:15] :stresnet.utils.logger: [INFO] loss: 17.55829429626465
-[2022-08-21 13:33:57] :stresnet.utils.logger: [INFO] loss: 17.338008880615234
-[2022-08-21 13:34:39] :stresnet.utils.logger: [INFO] loss: 17.667972564697266
-[2022-08-21 13:35:20] :stresnet.utils.logger: [INFO] loss: 16.833003997802734
-[2022-08-21 13:36:02] :stresnet.utils.logger: [INFO] loss: 16.35778045654297
-[2022-08-21 13:36:44] :stresnet.utils.logger: [INFO] loss: 16.13718032836914
-[2022-08-21 13:37:26] :stresnet.utils.logger: [INFO] loss: 15.221627235412598
-[2022-08-21 13:38:07] :stresnet.utils.logger: [INFO] loss: 14.816903114318848
-[2022-08-21 13:38:49] :stresnet.utils.logger: [INFO] loss: 15.053122520446777
-[2022-08-21 13:39:31] :stresnet.utils.logger: [INFO] loss: 15.38190746307373
-[2022-08-21 13:40:12] :stresnet.utils.logger: [INFO] loss: 15.329069137573242
-[2022-08-21 13:40:54] :stresnet.utils.logger: [INFO] loss: 15.661240577697754
-[2022-08-21 13:41:36] :stresnet.utils.logger: [INFO] loss: 15.622575759887695
-[2022-08-21 13:42:18] :stresnet.utils.logger: [INFO] loss: 15.700996398925781
-[2022-08-21 13:42:59] :stresnet.utils.logger: [INFO] loss: 15.06043529510498
-[2022-08-21 13:43:41] :stresnet.utils.logger: [INFO] loss: 14.50886344909668
-[2022-08-21 13:44:23] :stresnet.utils.logger: [INFO] loss: 14.183765411376953
-[2022-08-21 13:45:05] :stresnet.utils.logger: [INFO] loss: 14.036737442016602
-[2022-08-21 13:45:46] :stresnet.utils.logger: [INFO] loss: 13.877324104309082
-[2022-08-21 13:46:28] :stresnet.utils.logger: [INFO] loss: 13.933774948120117
-[2022-08-21 13:47:10] :stresnet.utils.logger: [INFO] loss: 14.287158966064453
-[2022-08-21 13:47:52] :stresnet.utils.logger: [INFO] loss: 14.366249084472656
-[2022-08-21 13:48:33] :stresnet.utils.logger: [INFO] loss: 14.488205909729004
-[2022-08-21 13:49:15] :stresnet.utils.logger: [INFO] loss: 14.525459289550781
-[2022-08-21 13:49:57] :stresnet.utils.logger: [INFO] loss: 14.224540710449219
-[2022-08-21 13:50:38] :stresnet.utils.logger: [INFO] loss: 14.051094055175781
-[2022-08-21 13:51:20] :stresnet.utils.logger: [INFO] loss: 14.377730369567871
-[2022-08-21 13:52:02] :stresnet.utils.logger: [INFO] loss: 15.139881134033203
-[2022-08-21 13:52:44] :stresnet.utils.logger: [INFO] loss: 15.4631929397583
-[2022-08-21 13:53:25] :stresnet.utils.logger: [INFO] loss: 15.793074607849121
-[2022-08-21 13:54:07] :stresnet.utils.logger: [INFO] loss: 16.54371452331543
-[2022-08-21 13:54:49] :stresnet.utils.logger: [INFO] loss: 15.94629192352295
-[2022-08-21 13:55:30] :stresnet.utils.logger: [INFO] loss: 15.96289348602295
-[2022-08-21 13:56:12] :stresnet.utils.logger: [INFO] loss: 15.459577560424805
-[2022-08-21 13:56:54] :stresnet.utils.logger: [INFO] loss: 14.718029022216797
-[2022-08-21 13:57:36] :stresnet.utils.logger: [INFO] loss: 14.719179153442383
-[2022-08-21 13:58:17] :stresnet.utils.logger: [INFO] loss: 14.359094619750977
-[2022-08-21 13:58:59] :stresnet.utils.logger: [INFO] loss: 13.813424110412598
-[2022-08-21 13:59:41] :stresnet.utils.logger: [INFO] loss: 12.542519569396973
-[2022-08-21 14:00:23] :stresnet.utils.logger: [INFO] loss: 12.069497108459473
-[2022-08-21 14:01:04] :stresnet.utils.logger: [INFO] loss: 12.010600090026855
-[2022-08-21 14:01:46] :stresnet.utils.logger: [INFO] loss: 12.745259284973145
-[2022-08-21 14:02:28] :stresnet.utils.logger: [INFO] loss: 13.263908386230469
-[2022-08-21 14:03:10] :stresnet.utils.logger: [INFO] loss: 13.164176940917969
-[2022-08-21 14:03:51] :stresnet.utils.logger: [INFO] loss: 12.950017929077148
-[2022-08-21 14:04:33] :stresnet.utils.logger: [INFO] loss: 12.656583786010742
-[2022-08-21 14:05:15] :stresnet.utils.logger: [INFO] loss: 12.628523826599121
-[2022-08-21 14:05:56] :stresnet.utils.logger: [INFO] loss: 12.597163200378418
-[2022-08-21 14:06:38] :stresnet.utils.logger: [INFO] loss: 12.720017433166504
-[2022-08-21 14:07:20] :stresnet.utils.logger: [INFO] loss: 12.926918983459473
-[2022-08-21 14:08:02] :stresnet.utils.logger: [INFO] loss: 12.84497356414795
-[2022-08-21 14:08:43] :stresnet.utils.logger: [INFO] loss: 12.754738807678223
-[2022-08-21 14:09:25] :stresnet.utils.logger: [INFO] loss: 12.975921630859375
-[2022-08-21 14:10:07] :stresnet.utils.logger: [INFO] loss: 13.264533996582031
-[2022-08-21 14:10:49] :stresnet.utils.logger: [INFO] loss: 12.663485527038574
-[2022-08-21 14:11:30] :stresnet.utils.logger: [INFO] loss: 12.791598320007324
-[2022-08-21 14:12:12] :stresnet.utils.logger: [INFO] loss: 12.90368366241455
-[2022-08-21 14:12:54] :stresnet.utils.logger: [INFO] loss: 13.146836280822754
-[2022-08-21 14:13:36] :stresnet.utils.logger: [INFO] loss: 13.102580070495605
-[2022-08-21 14:14:17] :stresnet.utils.logger: [INFO] loss: 13.066079139709473
-[2022-08-21 14:14:59] :stresnet.utils.logger: [INFO] loss: 12.347060203552246
-[2022-08-21 14:15:41] :stresnet.utils.logger: [INFO] loss: 12.890955924987793
-[2022-08-21 14:16:23] :stresnet.utils.logger: [INFO] loss: 12.878378868103027
-[2022-08-21 14:17:04] :stresnet.utils.logger: [INFO] loss: 12.108634948730469
-[2022-08-21 14:17:46] :stresnet.utils.logger: [INFO] loss: 13.61639404296875
-[2022-08-21 14:18:28] :stresnet.utils.logger: [INFO] loss: 12.120988845825195
-[2022-08-21 14:19:09] :stresnet.utils.logger: [INFO] loss: 13.806083679199219
-[2022-08-21 14:19:51] :stresnet.utils.logger: [INFO] loss: 14.319558143615723
-[2022-08-21 14:20:33] :stresnet.utils.logger: [INFO] loss: 13.57078742980957
-[2022-08-21 14:21:15] :stresnet.utils.logger: [INFO] loss: 13.881075859069824
