+++ examples/TaxiBJ/L6-C3-P1-T1/config.ini
@@ -1,0 +1,24 @@
+[dataset]
+data_files = ["./datasets/TaxiBJ/BJ13_M32x32_T30_InOut.h5", "./datasets/TaxiBJ/BJ14_M32x32_T30_InOut.h5", "./datasets/TaxiBJ/BJ15_M32x32_T30_InOut.h5", "./datasets/TaxiBJ/BJ16_M32x32_T30_InOut.h5"]
+holiday_file = ./datasets/TaxiBJ/BJ_Holiday.txt
+meteorol_file = ./datasets/TaxiBJ/BJ_Meteorology.h5
+T = 48
+len_closeness = 3
+len_period = 1
+len_trend = 1
+period_interval = 1
+trend_interval = 7
+; 48 * 28
+len_test = 1344
+use_meta = true
+
+[model]
+nb_flow = 2
+map_height = 32
+map_width = 32
+nb_residual_unit = 6
+
+[learning]
+epochs = 100
+batch_size = 32
+learning_rate = 0.0002
