+++ stresnet/utils/cfg.py
@@ -1,3 +1,4 @@
+import json
 from configparser import ConfigParser
 from dataclasses import dataclass
 
@@ -21,6 +22,10 @@ class Cfg:
 
     def _dataset_params(self):
         return {
+            "data_files": json.loads(self.config.get("dataset", "data_files")),
+            "holiday_file": self.config.get("dataset", "holiday_file"),
+            "meteorol_file": self.config.get("dataset", "meteorol_file"),
+            "T": self.config.getint("dataset", "T"),
             "len_closeness": self.config.getint("dataset", "len_closeness"),
             "len_period": self.config.getint("dataset", "len_period"),
             "len_trend": self.config.getint("dataset", "len_trend"),
@@ -28,8 +33,6 @@ class Cfg:
             "trend_interval": self.config.getint("dataset", "trend_interval"),
             "len_test": self.config.getint("dataset", "len_test"),
             "use_meta": self.config.getboolean("dataset", "use_meta"),
-            "use_holiday": self.config.getboolean("dataset", "use_holiday"),
-            "use_meteorol": self.config.getboolean("dataset", "use_meteorol"),
         }
 
     def _model_params(self):
